Admin previliges
Added Admin attribute to the users
Only the admin can see the project details, assignee details and edit the assignee profile and data.
Added constraints on the views (the show button) and the routes for those specific pages.
Amend: added the same constraints on the import / delete buttons for the projects and the assignees.